refactor: create test folder
- test 관련 모듈 모아놓을 test 폴더 생성
- test에 필요한 데이터 testingData로 이동
